Fixed dashboard a bit
