+++ frontend/src/app/layout.tsx
@@ -1,13 +1,15 @@
-import type { Metadata } from "next";
-import { Inter } from "next/font/google";
-import "./globals.css";
-import Navbar from "@/components/Navbar";
+// RootLayout.tsx
+import React from 'react';
+import type { Metadata } from 'next';
+import { Inter } from 'next/font/google';
+import './globals.css'; // Ensure this file includes Tailwind imports
+import Navbar from '@/components/Navbar'; // Adjust import path as necessary
 
-const inter = Inter({ subsets: ["latin"] });
+const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: "FlightBook",
-  description: "Generated by create next app",
+  title: 'FlightBook',
+  description: 'Generated by create next app',
 };
 
 export default function RootLayout({
@@ -17,8 +19,12 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <Navbar />
-      <body className={inter.className}>{children}</body>
+      <body className={`${inter.className} bg-gray-100`}>
+        <Navbar />
+        <div className="pt-16">
+          {children}
+        </div>
+      </body>
     </html>
   );
 }
